backend/libraryproject/requirements.txt
asgiref==3.10.0
certifi==2025.11.12
cloudinary==1.44.1
Django==5.2.8
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
